Added Escaping and MySQLi Interaction
Added names to the form fields so that they can be used as an array. From there, I added the MySQLi database connection so that mysqli->real_escape_string($string) could be used. This is to prevent SQL injection. We also added the foreach loop to get the scholarships that were selected in the form.